+++ pkg/vault/contracts/VaultCommon.sol
@@ -191,11 +191,11 @@ abstract contract VaultCommon is IVaultEvents, IVaultErrors, VaultStorage, Reent
     }
 
     /// @dev Lowest level routine that plucks only the minimum necessary parts from storage.
-    function _getPoolPausedState(address pool) internal view returns (bool, uint256) {
+    function _getPoolPausedState(address pool) internal view returns (bool, uint32) {
         PoolConfigBits memory config = _poolConfig[pool];
 
         bool isPoolPaused = config.isPoolPaused();
-        uint256 pauseWindowEndTime = config.getPauseWindowEndTime();
+        uint32 pauseWindowEndTime = config.getPauseWindowEndTime();
 
         // Use the Vault's buffer period.
         // solhint-disable-next-line not-rely-on-time
