+++ pkg/vault/contracts/test/PoolFactoryMock.sol
@@ -17,7 +17,7 @@ contract PoolFactoryMock is FactoryWidePauseWindow {
 
     IVault private immutable _vault;
 
-    constructor(IVault vault, uint256 pauseWindowDuration) FactoryWidePauseWindow(pauseWindowDuration) {
+    constructor(IVault vault, uint32 pauseWindowDuration) FactoryWidePauseWindow(pauseWindowDuration) {
         _vault = vault;
     }
 
@@ -88,7 +88,7 @@ contract PoolFactoryMock is FactoryWidePauseWindow {
         address pool,
         TokenConfig[] memory tokenConfig,
         uint256 swapFee,
-        uint256 pauseWindowDuration,
+        uint32 pauseWindowDuration,
         bool protocolFeeExempt,
         PoolRoleAccounts memory roleAccounts,
         address poolHooksContract
@@ -97,7 +97,7 @@ contract PoolFactoryMock is FactoryWidePauseWindow {
             pool,
             tokenConfig,
             swapFee,
-            block.timestamp + pauseWindowDuration,
+            uint32(block.timestamp) + pauseWindowDuration,
             protocolFeeExempt,
             roleAccounts,
             poolHooksContract,
@@ -153,7 +153,7 @@ contract PoolFactoryMock is FactoryWidePauseWindow {
     function registerPoolAtTimestamp(
         address pool,
         TokenConfig[] memory tokenConfig,
-        uint256 timestamp,
+        uint32 timestamp,
         PoolRoleAccounts memory roleAccounts,
         address poolHooksContract,
         LiquidityManagement calldata liquidityManagement
