+++ pkg/vault/test/foundry/unit/PoolAndVaultPaused.t.sol
@@ -15,7 +15,7 @@ import { BaseVaultTest } from "../utils/BaseVaultTest.sol";
 contract PoolAndVaultPausedTest is BaseVaultTest {
     // A number that is much smaller than the vault pause buffer end time, so we can play with
     // pool and vault pause windows.
-    uint256 private constant _FIXED_POOL_PAUSE_END_TIME = 1e5;
+    uint32 private constant _FIXED_POOL_PAUSE_END_TIME = 1e5;
     uint256 private _vaultBufferPeriodEndTimeTest;
 
     function setUp() public virtual override {
